Oops! Accidentally uploaded Parse API tokens
I previously refactored the API tokens to a .plist file which is read
in the application delegate, but I had forgotten to remove the
commented-out call to Parse.setApplicationId which exposed those tokens
in plain text.

I had to delete the app on Parse and recreate it to get new API tokens.
Oh well.