Support for user authentication
Users can be signed up, logged in, and have passwords reset.
Default ACL for Actions created is limited to read/write for the user
that created it, and does not include public read access.
Each Action has a ‘user’ property which relates it to the PFUser that
created it.
Actions can be created with an automatic anonymous user if no user is
authenticated.
Anonymous users are converted when the sign up workflow is used, and
associated Actions are automatically associated with the converted
authenticated user.